// frontend/src/app/utils/enums/permissions.enum.ts
export enum PermissionsEnum {
  DASHBOARD = "DASHBOARD",
  CONSULTATION_RECORD = "CONSULTATION_RECORD",
  APPOINTMENT_MANAGEMENT = "APPOINTMENT_MANAGEMENT",
  SUPPLIES = "SUPPLIES",
  FINANCE = "FINANCE",
  REPORTS = "REPORTS",
  CONFIGURATION = "CONFIGURATION",
  SYSTEM = "SYSTEM",
  PATIENTS = "PATIENTS",
}

export enum ActionsEnum {
  READ = "READ",
  CREATE = "CREATE",
  DELETE = "DELETE",
  UPDATE = "UPDATE",
  UPLOAD = "UPLOAD",
}